+++ sw.js
@@ -34,6 +34,7 @@ const urlsToCache = [
     './js/modules/notifications.js',
     './js/modules/clipboardManager.js',
     './js/modules/keyboard.js',
+    './js/modules/permissions.js',
     './js/apps/appStore/main.js',
     './js/apps/browser/main.js',
     './js/apps/calculator/main.js',